Handle error when checking if kaggle data already downloaded

+++ ml-utils/ml_utils/kaggle_utils/__init__.py
@@ -6,7 +6,10 @@ from pathlib import Path
 
 def _data_already_exists() -> bool:
     data_dir = Path("data")
-    downloaded_files = os.listdir(data_dir)
+    try:
+        downloaded_files = os.listdir(data_dir)
+    except FileNotFoundError:
+        return False
     return len(downloaded_files) > 0
 
 
@@ -29,6 +32,10 @@ def download_and_unzip(competition: str) -> None:
         "-c",
         competition
     ])
+
+    if not zip_file_name.exists():
+        raise FileNotFoundError("Kaggle failed to download, check local kaggle.json")
+
     data_dir.mkdir(exist_ok=True)
     shutil.unpack_archive(str(zip_file_name), data_dir)
     Path(zip_file_name).unlink()
